Add '/profile' route to RESTful API
Introduce a new '/profile' route in the API to access the profile resource. Following RESTful principles, this route allows for read, update, and delete operations on the profile. The server is configured to provide profile information in JSON format in response to a GET request. This approach facilitates structured interaction with the profile resource through the HTTP protocol.