Initiate XMLHttpRequest for fetching profile data
Create an instance of XMLHttpRequest to enable making HTTP requests from the client. Open an HTTP connection using the GET method to the 'http://localhost:3000/profile' URL. Send the request to the server, initiating communication to retrieve profile information. Set up an event handler to execute code when the request is successfully completed, obtaining the server's response in text format, which may be JSON or another format.